Move vendor to _internal
