+++ .pre-commit-config.yaml
@@ -9,12 +9,12 @@ repos:
     rev: 22.10.0
     hooks:
     - id: black
-      exclude: tanjun/_vendor
+      exclude: tanjun/_internal/vendor
   - repo: https://github.com/PyCQA/isort
     rev: 5.10.1
     hooks:
     - id: isort
-      exclude: tanjun/_vendor
+      exclude: tanjun/_internal/vendor
   - repo: https://github.com/aio-libs/sort-all
     rev: v1.2.0
     hooks:
