+++ tanjun/_internal/__init__.py
@@ -46,7 +46,7 @@ from collections import abc as collections
 import hikari
 
 from .. import errors
-from .._vendor import inspect
+from .._internal.vendor import inspect
 
 if typing.TYPE_CHECKING:
     import typing_extensions
